fix(build): overlay page の不正な named export で next build が失敗する問題を修正
Next.js の Page ファイルは規定フィールド以外の export を持てず、#595 で
page.tsx に追加した pickSoundBearingCardIndex の named export が
next build の Page 型検証で失敗していた(vitest/tsc では検出されない
build 固有の検査。preview の Workers Builds fail も同原因)。
関数を gacha-sound-rules.ts(同じ優先度定義を持つ pickGachaSoundRule の隣)へ
移動し、page/テストの import を追随。ロジックは無変更。
